Fix multiple checkins
There was a previous issue whereby users could checkin more than
once to a single pub. This fix makes it so that existence of a checkin
is verified before a new checkin is added. This is more of a quickfix
and will need to be updated when the route store is moved.